requirements.txt
# YOLOv8 训练系统依赖

# 核心依赖
ultralytics>=8.0.0
flask>=3.0.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0.0
numpy>=1.24.0

# 深度学习框架（ultralytics会自动安装torch，但也可以指定版本）
torch>=2.0.0
torchvision>=0.15.0

# API支持
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# 工具库
werkzeug>=3.0.0

# 打包工具
pyinstaller>=6.0.0
pywebview>=4.0.0  # 桌面客户端支持

# 可选（加速训练）
# tensorboard  # 训练可视化
# onnx         # 模型导出
# onnxruntime  # ONNX推理
